refactor(DashboardLayout): simplify layout structure and remove unused components
